Change to two settings tabs
