+++ src/components/settings/AccountSettings/AccountSettings.tsx
@@ -1,0 +1,5 @@
+const AccountSettings = () => {
+  return <div>AccountSettings</div>;
+};
+
+export default AccountSettings;
